requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.5
aiosignal==1.3.1; python_version >= '3.7'
async-timeout==4.0.2; python_version >= '3.6'
attrs==23.1.0; python_version >= '3.7'
beautifulsoup4==4.12.2
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
discord.py==2.3.1
frozenlist==1.4.0; python_version >= '3.8'
idna==3.4; python_version >= '3.5'
multidict==6.0.4; python_version >= '3.7'
soupsieve==2.4.1; python_version >= '3.7'
typing==3.7.4.3
yarl==1.9.2; python_version >= '3.7'
